move cache code to worker + to types.d.ts
