+++ packages/components/src/workers/parquetWorker.ts
@@ -1,14 +1,14 @@
-import { ColumnData, parquetQuery } from 'hyparquet'
+import { AsyncBuffer, ColumnData, parquetQuery } from 'hyparquet'
 import { compressors } from 'hyparquet-compressors'
-import {
+import { asyncBufferFromUrl } from '../lib/utils.ts'
+import type {
+  AsyncBufferFrom,
   ChunkMessage,
   ErrorMessage,
   IndicesMessage,
   ParquetReadWorkerOptions,
   ResultMessage,
-  asyncBufferFrom,
-  compare,
-} from './parquetWorkerClient.ts'
+} from './types.d.ts'
 
 function postChunkMessage ({ chunk, queryId }: ChunkMessage) {
   self.postMessage({ chunk, queryId })
@@ -22,6 +22,9 @@ function postErrorMessage ({ error, queryId }: ErrorMessage) {
 function postIndicesMessage ({ indices, queryId }: IndicesMessage) {
   self.postMessage({ indices, queryId })
 }
+
+// Cache for AsyncBuffers
+const cache = new Map<string, Promise<AsyncBuffer>>()
 
 self.onmessage = async ({
   data,
@@ -81,5 +84,79 @@ self.onmessage = async ({
     } catch (error) {
       postErrorMessage({ error: error as Error, queryId })
     }
+  }
+}
+
+function compare<T>(a: T, b: T): number {
+  if (a < b) return -1
+  if (a > b) return 1
+  return 1 // TODO: how to handle nulls?
+}
+
+/**
+ * Convert AsyncBufferFrom to AsyncBuffer and cache results.
+ */
+function asyncBufferFrom(
+  from: AsyncBufferFrom,
+): Promise<AsyncBuffer> {
+  const key = JSON.stringify(from)
+  const cached = cache.get(key)
+  if (cached) return cached
+  const asyncBuffer = asyncBufferFromUrl(from).then(cachedAsyncBuffer)
+  cache.set(key, asyncBuffer)
+  return asyncBuffer
+}
+
+type Awaitable<T> = T | Promise<T>;
+
+function cachedAsyncBuffer(asyncBuffer: AsyncBuffer): AsyncBuffer {
+  const cache = new Map<string, Awaitable<ArrayBuffer>>()
+  const { byteLength } = asyncBuffer
+  return {
+    byteLength,
+    /**
+     * @param {number} start
+     * @param {number} [end]
+     * @returns {Awaitable<ArrayBuffer>}
+     */
+    slice(start: number, end?: number): Awaitable<ArrayBuffer> {
+      const key = cacheKey(start, end, byteLength)
+      const cached = cache.get(key)
+      if (cached) return cached
+      // cache miss, read from file
+      const promise = asyncBuffer.slice(start, end)
+      cache.set(key, promise)
+      return promise
+    },
+  }
+}
+
+/**
+ * Returns canonical cache key for a byte range 'start,end'.
+ * Normalize int-range and suffix-range requests to the same key.
+ *
+ * @param {number} start start byte of range
+ * @param {number} [end] end byte of range, or undefined for suffix range
+ * @param {number} [size] size of file, or undefined for suffix range
+ * @returns {string}
+ */
+function cacheKey(start: number, end?: number, size?: number): string {
+  if (start < 0) {
+    if (end !== undefined)
+      throw new Error(
+        `invalid suffix range [${start.toString()}, ${end.toString()}]`,
+      )
+    if (size === undefined) return `${start.toString()},`
+    return `${(size + start).toString()},${size.toString()}`
+  } else if (end !== undefined) {
+    if (start > end)
+      throw new Error(
+        `invalid empty range [${start.toString()}, ${end.toString()}]`,
+      )
+    return `${start.toString()},${end.toString()}`
+  } else if (size === undefined) {
+    return `${start.toString()},`
+  } else {
+    return `${start.toString()},${size.toString()}`
   }
 }
